Add drag and drop functionality
This commits adds the start of our drag and drop functionality. Once
finished the user will be able to drag a song to a playlist to add it to
that playlist. This commit also tightens up some of our routes by
redirecting root to the playlists path. A settings page was added that
we can work off of as well.