add-infinite-scroll
Fixes the 50th duplicate song bug. I think this is fine as far as 
infinite scroll goes. I considered us using it for the search section, 
but I feel like that might be a little too much information for one 
page, so I changed the search limit to 30. If the user doesn't find the 
song they're searching for, then I think they'll just have to be more 
specific in the search section rather than it loading every possible 
song.

Otherwise, thanks for helping me out.